+++ demo/src/main/java/org/chzz/refresh/demo/ui/activity/NormalRecyclerViewActivity.java
@@ -9,6 +9,15 @@ import android.widget.ImageView;
 
 import com.bumptech.glide.Glide;
 
+import org.chzz.refresh.CHZZMoocStyleRefreshViewHolder;
+import org.chzz.refresh.CHZZRefreshLayout;
+import org.chzz.refresh.demo.R;
+import org.chzz.refresh.demo.adapter.NormalRecyclerViewAdapter;
+import org.chzz.refresh.demo.model.BannerModel;
+import org.chzz.refresh.demo.model.RefreshModel;
+import org.chzz.refresh.demo.util.ThreadUtil;
+import org.chzz.refresh.demo.widget.Divider;
+
 import java.util.ArrayList;
 import java.util.List;
 
@@ -17,21 +26,10 @@ import cn.bingoogolapple.androidcommon.adapter.BGAOnItemChildLongClickListener;
 import cn.bingoogolapple.androidcommon.adapter.BGAOnRVItemClickListener;
 import cn.bingoogolapple.androidcommon.adapter.BGAOnRVItemLongClickListener;
 import cn.bingoogolapple.bgabanner.BGABanner;
-
-import org.chzz.refresh.demo.adapter.NormalRecyclerViewAdapter;
-import org.chzz.refresh.demo.model.BannerModel;
-import org.chzz.refresh.demo.model.RefreshModel;
-import org.chzz.refresh.demo.util.ThreadUtil;
-import org.chzz.refresh.CHZZMoocStyleRefreshViewHolder;
-import org.chzz.refresh.CHZZRefreshLayout;
-
-import org.chzz.refresh.demo.R;
-
-import org.chzz.refresh.demo.widget.Divider;
 import retrofit2.Callback;
 import retrofit2.Response;
 
-public class NormalRecyclerViewActivity extends BaseActivity implements BGAOnRVItemClickListener, BGAOnRVItemLongClickListener, BGAOnItemChildClickListener, BGAOnItemChildLongClickListener, CHZZRefreshLayout.BGARefreshLayoutDelegate {
+public class NormalRecyclerViewActivity extends BaseActivity implements BGAOnRVItemClickListener, BGAOnRVItemLongClickListener, BGAOnItemChildClickListener, BGAOnItemChildLongClickListener, CHZZRefreshLayout.RefreshLayoutDelegate {
     private CHZZRefreshLayout mRefreshLayout;
     private BGABanner mBanner;
     private RecyclerView mDataRv;
@@ -150,7 +148,7 @@ public class NormalRecyclerViewActivity extends BaseActivity implements BGAOnRVI
     }
 
     @Override
-    public void onBGARefreshLayoutBeginRefreshing(CHZZRefreshLayout refreshLayout) {
+    public void onRefreshLayoutBeginRefreshing(CHZZRefreshLayout refreshLayout) {
         mNewPageNumber++;
         if (mNewPageNumber > 4) {
             mRefreshLayout.endRefreshing();
@@ -182,7 +180,7 @@ public class NormalRecyclerViewActivity extends BaseActivity implements BGAOnRVI
     }
 
     @Override
-    public boolean onBGARefreshLayoutBeginLoadingMore(CHZZRefreshLayout refreshLayout) {
+    public boolean onRefreshLayoutBeginLoadingMore(CHZZRefreshLayout refreshLayout) {
         mMorePageNumber++;
         if (mMorePageNumber > 4) {
             mRefreshLayout.endLoadingMore();