+++ demo/src/main/java/org/chzz/refresh/demo/ui/fragment/StickyNavWebViewFragment.java
@@ -15,7 +15,7 @@ import org.chzz.refresh.demo.ui.activity.ViewPagerActivity;
  * 创建时间:15/9/27 下午12:53
  * 描述:
  */
-public class StickyNavWebViewFragment extends BaseFragment implements CHZZRefreshLayout.BGARefreshLayoutDelegate {
+public class StickyNavWebViewFragment extends BaseFragment implements CHZZRefreshLayout.RefreshLayoutDelegate {
     private WebView mContentWv;
 
     @Override
@@ -51,12 +51,12 @@ public class StickyNavWebViewFragment extends BaseFragment implements CHZZRefres
     }
 
     @Override
-    public void onBGARefreshLayoutBeginRefreshing(CHZZRefreshLayout refreshLayout) {
+    public void onRefreshLayoutBeginRefreshing(CHZZRefreshLayout refreshLayout) {
         mContentWv.reload();
     }
 
     @Override
-    public boolean onBGARefreshLayoutBeginLoadingMore(CHZZRefreshLayout refreshLayout) {
+    public boolean onRefreshLayoutBeginLoadingMore(CHZZRefreshLayout refreshLayout) {
         Log.i(TAG, "加载更多");
         return false;
     }